Replaced notifications UI test with the FCM version
