--- test-application/app/styles/fcmnotifications.tss
@@ -1,10 +1,0 @@
-
-".container" : {
-
-}
-
-"#tokenLabel" : {
-	font: {
-		fontSize: 11,
-	}
-}
