+++ test-application/app/styles/notifications.tss
@@ -1,4 +1,10 @@
 
 ".container" : {
-  
+
+}
+
+"#tokenLabel" : {
+	font: {
+		fontSize: 11,
+	}
 }
