+++ test-application/app/views/notifications.xml
@@ -1,9 +1,11 @@
 <Alloy>
 	<Window id="win" layout="vertical" title="Notifications">
 		<View height="40" />
+		<Label text="Inactive" id="activeLbl" color="red" />
 		<Button title="Activate" id="activateBtn" top="10" />
-		<Button title="Deactivate" id="deactivateBtn" top="10" />
 		<Button title="Subscribe" id="subscribeBtn" top="10" />
 		<Button title="Unsubscribe" id="unsubscribeBtn" top="10" />
+		<Label text="FCM Token:" id="tokenTitle" top="16" />
+		<Label text="" id="tokenLabel" top="8" left="32" right="32" />
 	</Window>
 </Alloy>
